Improved syntax
